Return type fix on fibonachi

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -1,6 +1,6 @@
 #include <stdio.h>
 
-long main(void)
+int main(void)
 {
 int i;
 long t1 = 0, t2 = 1;
@@ -11,18 +11,18 @@ for (i = 2; i < 51; i++)
 if (i <= 1)
 {
 next = i;
-printf("%d, ", next);
+printf("%lu, ", next);
 }
 else if (i < 51)
 {
 next = t1 + t2;
 t1 = t2;
 t2 = next;
-printf("%d, ", next);
+printf("%lu, ", next);
 }
 else
 {
-printf("%d", next);
+printf("%lu", next);
 }
 printf("\n");
 return (0);
